refactor(agents): keep stored override normalization bounded
